sqlalchemy and structure
