+++ Dockerfile
@@ -7,4 +7,4 @@ RUN pip3 install -r requirements.txt
 
 COPY . .
 
-CMD [ "python3", "app.py"]
+CMD [ "python3", "run.py"]
