+++ core/models.py
@@ -1,0 +1,15 @@
+from core import app
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy(app)
+
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
+    name = db.Column(db.String(80), nullable=False)
+    surname = db.Column(db.String(80), nullable=False)
+    email = db.Column(db.String(120), unique=True, nullable=False)
+
+    def __init__(self, name, surname, email):
+        self.name = name
+        self.surname = surname
+        self.email = email
