+++ requirements.txt
@@ -4,6 +4,7 @@ charset-normalizer==2.0.7
 click==8.0.3
 Flask==2.0.2
 flask-ngrok==0.0.25
+Flask-SQLAlchemy==2.5.1
 idna==3.3
 itsdangerous==2.0.1
 Jinja2==3.0.3
@@ -12,5 +13,6 @@ MarkupSafe==2.0.1
 psycopg2-binary==2.9.2
 pyrsistent==0.18.0
 requests==2.26.0
+SQLAlchemy==1.4.29
 urllib3==1.26.7
 Werkzeug==2.0.2
